pyproject.toml
[tool.poetry]
name = "todos"
version = "0.1.0"
description = ""
authors = ["SeeraSidhu <serrasidhu0@gmail.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.12"

unicorn = {extras = ["strand"], version = "^2.1.1"}
fastapi = "^0.115.2"
uvicorn = "^0.32.0"
langchain = "^0.3.4"
python-dotenv = "^1.0.1"
langchain-google-genai = "^2.0.1"
[tool.poetry.scripts]
dev = "uvicorn todos.main:app --reload"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
